Move all top-bar tests into one file
