+++ src/renderer/components/layout/top-bar/top-bar.test.tsx
@@ -13,8 +13,6 @@ import type { DiRender } from "../../test-utils/renderFor";
 import { renderFor } from "../../test-utils/renderFor";
 import topBarItemsInjectable from "./top-bar-items/top-bar-items.injectable";
 import { computed, observable } from "mobx";
-import isLinuxInjectable from "../../../../common/vars/is-linux.injectable";
-import isWindowsInjectable from "../../../../common/vars/is-windows.injectable";
 import rendererExtensionsInjectable from "../../../../extensions/renderer-extensions.injectable";
 import closeWindowInjectable from "./close-window.injectable";
 import goBackInjectable from "./go-back.injectable";
@@ -23,23 +21,28 @@ import maximizeWindowInjectable from "./maximize-window.injectable";
 import openAppContextMenuInjectable from "./open-app-context-menu.injectable";
 import toggleMaximizeWindowInjectable from "./toggle-maximize-window.injectable";
 import topBarStateInjectable from "./state.injectable";
+import platformInjectable from "../../../../common/vars/platform.injectable";
 
 describe("<TopBar/>", () => {
   let di: DiContainer;
   let render: DiRender;
   let goBack: jest.MockedFunction<() => void>;
   let goForward: jest.MockedFunction<() => void>;
+  let openAppContextMenu: jest.MockedFunction<() => void>;
+  let closeWindow: jest.MockedFunction<() => void>;
+  let maximizeWindow: jest.MockedFunction<() => void>;
+  let toggleMaximizeWindow: jest.MockedFunction<() => void>;
 
   beforeEach(() => {
     di = getDiForUnitTesting({ doGeneralOverrides: true });
 
     di.override(rendererExtensionsInjectable, () => computed(() => []));
-    di.override(openAppContextMenuInjectable, () => jest.fn());
+    di.override(openAppContextMenuInjectable, () => openAppContextMenu = jest.fn());
     di.override(goBackInjectable, () => goBack = jest.fn());
     di.override(goForwardInjectable, () => goForward = jest.fn());
-    di.override(closeWindowInjectable, () => jest.fn());
-    di.override(maximizeWindowInjectable, () => jest.fn());
-    di.override(toggleMaximizeWindowInjectable, () => jest.fn());
+    di.override(closeWindowInjectable, () => closeWindow = jest.fn());
+    di.override(maximizeWindowInjectable, () => maximizeWindow = jest.fn());
+    di.override(toggleMaximizeWindowInjectable, () => toggleMaximizeWindow = jest.fn());
 
     render = renderFor(di);
   });
@@ -109,28 +112,91 @@ describe("<TopBar/>", () => {
       expect(await findByTestId(testId)).toHaveTextContent(text);
     });
 
-    it("doesn't show windows title buttons on macos", () => {
-      di.override(isLinuxInjectable, () => false);
-      di.override(isWindowsInjectable, () => false);
+    describe("on macos", () => {
+      beforeEach(() => {
+        di.override(platformInjectable, () => "darwin");
+      });
 
-      const { queryByTestId } = render(<TopBar/>);
+      it("doesn't show windows title", () => {
+        const { queryByTestId } = render(<TopBar/>);
 
-      expect(queryByTestId("window-menu")).not.toBeInTheDocument();
-      expect(queryByTestId("window-minimize")).not.toBeInTheDocument();
-      expect(queryByTestId("window-maximize")).not.toBeInTheDocument();
-      expect(queryByTestId("window-close")).not.toBeInTheDocument();
+        expect(queryByTestId("window-menu")).not.toBeInTheDocument();
+        expect(queryByTestId("window-minimize")).not.toBeInTheDocument();
+        expect(queryByTestId("window-maximize")).not.toBeInTheDocument();
+        expect(queryByTestId("window-close")).not.toBeInTheDocument();
+      });
     });
 
-    it("does show windows title buttons on linux", () => {
-      di.override(isLinuxInjectable, () => true);
-      di.override(isWindowsInjectable, () => false);
+    describe("on linux", () => {
+      beforeEach(() => {
+        di.override(platformInjectable, () => "linux");
+      });
 
-      const { queryByTestId } = render(<TopBar/>);
+      it("does show windows title buttons", () => {
+        const { queryByTestId } = render(<TopBar/>);
 
-      expect(queryByTestId("window-menu")).toBeInTheDocument();
-      expect(queryByTestId("window-minimize")).toBeInTheDocument();
-      expect(queryByTestId("window-maximize")).toBeInTheDocument();
-      expect(queryByTestId("window-close")).toBeInTheDocument();
+        expect(queryByTestId("window-menu")).toBeInTheDocument();
+        expect(queryByTestId("window-minimize")).toBeInTheDocument();
+        expect(queryByTestId("window-maximize")).toBeInTheDocument();
+        expect(queryByTestId("window-close")).toBeInTheDocument();
+      });
+
+      it("triggers ipc events on click", () => {
+        const { getByTestId } = render(<TopBar />);
+
+        const menu = getByTestId("window-menu");
+        const minimize = getByTestId("window-minimize");
+        const maximize = getByTestId("window-maximize");
+        const close = getByTestId("window-close");
+
+        fireEvent.click(menu);
+        expect(openAppContextMenu).toBeCalled();
+
+        fireEvent.click(minimize);
+        expect(maximizeWindow).toBeCalled();
+
+        fireEvent.click(maximize);
+        expect(toggleMaximizeWindow).toBeCalled();
+
+        fireEvent.click(close);
+        expect(closeWindow).toBeCalled();
+      });
+    });
+
+    describe("on windows", () => {
+      beforeEach(() => {
+        di.override(platformInjectable, () => "win32");
+      });
+
+      it("does show windows title buttons", () => {
+        const { queryByTestId } = render(<TopBar/>);
+
+        expect(queryByTestId("window-menu")).toBeInTheDocument();
+        expect(queryByTestId("window-minimize")).toBeInTheDocument();
+        expect(queryByTestId("window-maximize")).toBeInTheDocument();
+        expect(queryByTestId("window-close")).toBeInTheDocument();
+      });
+
+      it("triggers ipc events on click", () => {
+        const { getByTestId } = render(<TopBar />);
+
+        const menu = getByTestId("window-menu");
+        const minimize = getByTestId("window-minimize");
+        const maximize = getByTestId("window-maximize");
+        const close = getByTestId("window-close");
+
+        fireEvent.click(menu);
+        expect(openAppContextMenu).toBeCalled();
+
+        fireEvent.click(minimize);
+        expect(maximizeWindow).toBeCalled();
+
+        fireEvent.click(maximize);
+        expect(toggleMaximizeWindow).toBeCalled();
+
+        fireEvent.click(close);
+        expect(closeWindow).toBeCalled();
+      });
     });
   });
 });
